+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 FEATURES:
 
 - `sks_nodepool`: add `kubelet_max_pods` attribute to set the maximum number of pods per node on the underlying instance pool (resource + data sources)
+- dbaas: support for clickhouse
+- `dbaas`: add exoscale_dbaas_clickhouse_user resource
 
 ## 0.71.0